+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/NotificationsSettingsFragment.kt
@@ -71,6 +71,7 @@ class NotificationsSettingsFragment :
     private var enabledPreference: SwitchPreference? = null
     private var systemSettingsPreference: Preference? = null
     private var notificationActions: PreferenceScreen? = null
+    private var playOverNotificationPreference: ListPreference? = null
 
     private val toolbar
         get() = view?.findViewById<Toolbar>(R.id.toolbar)
@@ -95,6 +96,7 @@ class NotificationsSettingsFragment :
         vibratePreference = manager.findPreference("notificationVibrate")
         notificationActions = manager.findPreference("notificationActions")
         systemSettingsPreference = manager.findPreference("openSystemSettings")
+        playOverNotificationPreference = manager.findPreference("notificationAudio")
 
         // turn preferences off by default, because they are enable async, we don't want this view to remove them from the screen after it loads as it looks jarring
         enabledPreferences(false)
@@ -111,13 +113,6 @@ class NotificationsSettingsFragment :
         updateNotificationsEnabled()
 
         manager.run {
-            findPreference<SwitchPreference>(Settings.PREFERENCE_OVERRIDE_AUDIO)?.setOnPreferenceChangeListener { _, newValue ->
-                analyticsTracker.track(
-                    AnalyticsEvent.SETTINGS_NOTIFICATIONS_PLAY_OVER_NOTIFICATIONS_TOGGLED,
-                    mapOf("enabled" to newValue as Boolean)
-                )
-                true
-            }
             findPreference<SwitchPreference>(Settings.PREFERENCE_HIDE_NOTIFICATION_ON_PAUSE)?.setOnPreferenceChangeListener { _, newValue ->
                 analyticsTracker.track(
                     AnalyticsEvent.SETTINGS_NOTIFICATIONS_HIDE_PLAYBACK_NOTIFICATION_ON_PAUSE,
@@ -137,6 +132,13 @@ class NotificationsSettingsFragment :
                         else -> "unknown"
                     }
                 )
+            )
+            true
+        }
+        playOverNotificationPreference?.setOnPreferenceChangeListener { _, newValue ->
+            analyticsTracker.track(
+                AnalyticsEvent.SETTINGS_NOTIFICATIONS_PLAY_OVER_NOTIFICATIONS_TOGGLED,
+                mapOf("enabled" to (newValue != "2"))
             )
             true
         }
@@ -386,6 +388,7 @@ class NotificationsSettingsFragment :
     override fun onResume() {
         super.onResume()
         setupNotificationVibrate()
+        setupPlayOverNotifications()
         changePodcastsSummary()
         preferenceScreen.sharedPreferences?.registerOnSharedPreferenceChangeListener(this)
     }
@@ -398,6 +401,8 @@ class NotificationsSettingsFragment :
     override fun onSharedPreferenceChanged(sharedPreferences: SharedPreferences, key: String) {
         if (Settings.PREFERENCE_NOTIFICATION_VIBRATE == key) {
             changeVibrateSummary()
+        } else if (Settings.PREFERENCE_NOTIFICATION_AUDIO == key) {
+            changePlayOverNotificationSummary()
         }
     }
 
@@ -449,6 +454,28 @@ class NotificationsSettingsFragment :
             )
             it.entryValues = arrayOf("2", "1", "0")
             it.value = settings.getNotificationVibrate().toString()
+        }
+    }
+
+    private fun setupPlayOverNotifications() {
+        playOverNotificationPreference?.let {
+            it.entries = arrayOf(
+                getString(LR.string.settings_notification_play_over_never),
+                getString(LR.string.settings_notification_play_over_duck),
+                getString(LR.string.settings_notification_play_over_always)
+            )
+            it.entryValues = arrayOf("2", "1", "0")
+            it.value = settings.getPlayOverNotification().toString()
+        }
+        changePlayOverNotificationSummary()
+    }
+
+    private fun changePlayOverNotificationSummary() {
+        playOverNotificationPreference?.summary = when (settings.getPlayOverNotification()) {
+            2 -> getString(LR.string.settings_notification_play_over_never)
+            1 -> getString(LR.string.settings_notification_play_over_duck)
+            0 -> getString(LR.string.settings_notification_play_over_always)
+            else -> ""
         }
     }